+++ src/cosmos/cosmosInfos.ts
@@ -3,11 +3,13 @@ import { coreum } from './info/coreumInfo'
 import { cosmoshub } from './info/cosmoshubInfo'
 import { osmosis } from './info/osmosisInfo'
 import { thorchainrune } from './info/thorchainruneInfo'
+import { thorchainrunestagenet } from './info/thorchainruneStagenetInfo'
 
 export const cosmosPlugins = {
   axelar,
   coreum,
   cosmoshub,
   osmosis,
-  thorchainrune
+  thorchainrune,
+  thorchainrunestagenet
 }